Add SchemaErrorsItem and SchemaEventsItem to RefItem
